+++ packages/eslint-config-airbnb/test/test-react-order.js
@@ -1,6 +1,6 @@
 import test from 'tape';
 import { CLIEngine } from 'eslint';
-import eslintrc from '../';
+import eslintrc from '..';
 import reactRules from '../rules/react';
 import reactA11yRules from '../rules/react-a11y';
 
@@ -11,12 +11,14 @@ const cli = new CLIEngine({
   rules: {
     // It is okay to import devDependencies in tests.
     'import/no-extraneous-dependencies': [2, { devDependencies: true }],
+    // this doesn't matter for tests
+    'lines-between-class-members': 0,
   },
 });
 
 function lint(text) {
-  // @see http://eslint.org/docs/developer-guide/nodejs-api.html#executeonfiles
-  // @see http://eslint.org/docs/developer-guide/nodejs-api.html#executeontext
+  // @see https://eslint.org/docs/developer-guide/nodejs-api.html#executeonfiles
+  // @see https://eslint.org/docs/developer-guide/nodejs-api.html#executeontext
   const linter = cli.executeOnText(text);
   return linter.results[0];
 }
@@ -49,8 +51,7 @@ test('validate react prop order', (t) => {
   setBar() {}
   someMethod() {}
   renderDogs() {}
-  render() { return <div />; }
-`));
+  render() { return <div />; }`));
 
     t.notOk(result.warningCount, 'no warnings');
     t.notOk(result.errorCount, 'no errors');